requirements.txt
streamlit
langchain
langchain-community
langchain-core
langgraph
sentence-transformers
chromadb
pinecone-client
openai
tiktoken
python-dotenv
plotly
pandas
numpy
scikit-learn
umap-learn
requests
aiohttp
tqdm
pytest
black
flake8
mypy